Updated print messages

+++ awss3upload/src/main/java/example/S3Upload.java
@@ -1,9 +1,9 @@
 /**
- * S3Upload is an example that handles S3 buckets on AWS
- * Upload a local file to a S3 bucket
+ * S3Upload is an example that handles S3 buckets on AWS.
+ * Upload a local file to a S3 bucket.
  * You must provide 3 parameters:
  * BUCKET_NAME     = Bucket name
- * OBJECT_NAME     = Object file name in the bucket
+ * OBJECT_NAME     = Object name in the bucket
  * LOCAL_FILE_NAME = Local file name
  */
 
@@ -19,38 +19,42 @@ import com.amazonaws.services.s3.model.PutObjectRequest;
 
 public class S3Upload {
     public static void main(String[] args) throws IOException {
-        String bucketName;     // Bucket name
-        String keyName;        // Key name, it is the object name
-        String uploadFileName; // Upload local file name
+        String region = "eu-west-1";    // Region name
+        String bucketName;              // Bucket name
+        String keyName;                 // Key name, it is the object name
+        String localFileName;           // Upload local file name
 
         if (args.length < 3) {
-            System.out.println("Not enough parameters. Proper Usage is: java -jar s3upload.jar <BUCKET_NAME> <OBJECT_NAME> <LOCAL_FILE_NAME>");
+            System.out.println("Not enough parameters.\nProper Usage is: java -jar s3upload.jar <BUCKET_NAME> <OBJECT_NAME> <LOCAL_FILE_NAME>");
             System.exit(1);
         }
 
-        bucketName     = args[0];
-        keyName        = args[1];
-        uploadFileName = args[2];
+        bucketName    = args[0];
+        keyName       = args[1];
+        localFileName = args[2];
 
         System.out.println("Bucket:     " + bucketName);
         System.out.println("Object/Key: " + keyName);
-        System.out.println("Local file: " + uploadFileName);
+        System.out.println("Local file: " + localFileName);
 
         // Instantiates a client
-        AmazonS3 s3client = AmazonS3ClientBuilder.defaultClient();
+        AmazonS3 s3client = AmazonS3ClientBuilder.standard()
+                .withRegion(region).build();
 
         try {
             System.out.println("Uploading an object to S3 from a file ...");
 
             // Get local file
-            File file = new File(uploadFileName);
+            File file = new File(localFileName);
+            if (file.exists()) {
+                // Upload object
+                s3client.putObject(new PutObjectRequest(
+                        bucketName, keyName, file));
 
-            // Upload object
-            s3client.putObject(new PutObjectRequest(
-                    bucketName, keyName, file));
-
-            System.out.println("Uploaded");
-
+                System.out.println("Uploaded");
+            } else {
+                System.out.printf("Error: Local file \"%s\" does NOT exist.", localFileName);
+            }
         } catch (AmazonServiceException ase) {
             System.out.println("Caught an AmazonServiceException, " +
                     "which means your request made it " +
